requirements.txt
pypdf
pyyaml
numpy
sentence-transformers
streamlit
pypdfium2
ebooklib
beautifulsoup4
plotly
llama-cpp-python
fastapi
uvicorn[standard]
pydantic
langchain
langchain-core
langchain-community
django
djangorestframework
